+++ packages/angular/CHANGELOG.md
@@ -1,1 +1,5 @@
 # @agentskit/angular
+
+## Unreleased
+
+- Publish partial-Ivy output for production AOT consumers.